+++ CHANGELOG.md
@@ -34,6 +34,10 @@ VListRow
 
 - ListRow is deprecated in favor of more modern `List` APi
 
+VFetchingAsyncImage
+
+- A new component is moved from [VCore](https://github.com/VakhoKontridze/VCore), that asynchronously loads and displays an `Image` with a fetch handler
+
 ### [5.1.0(56)](https://github.com/VakhoKontridze/VComponents/releases/tag/5.1.0) — *2023 11 11*
 
 General